renaming src to apollo11_simulator, main to __main__
