+++ apollo11_simulator/__main__.py
@@ -1,5 +1,5 @@
-from src.models.event_processing.event_manager import EventManager
-from src.models.report_processing.report_builder import ReportBuilder
+from apollo11_simulator.models.event_processing.event_manager import EventManager
+from apollo11_simulator.models.report_processing.report_builder import ReportBuilder
 
 #logica argparse
 #si es en modo generador
