--- src/models/event_processing/vac_mars.py
@@ -1,13 +1,0 @@
-from src.models.event_processing.mission import Mission
-
-class VacMars(Mission):
-
-    number_of_passengers: int
-    ticket_price: float
-
-    @property
-    def total_sales(self) -> float:
-        return self.number_of_passengers * self.ticket_price
-    
-    def __str__(self) -> str:
-        return f'{super().__str__()}TMRS'
